restructure canary
