+++ go.work
@@ -3,5 +3,5 @@ go 1.18
 use (
 	.
 	./testground
-	./ops/aws/canary/src
+	./ops/aws/canary/lambda
 )
