+++ ops/aws/canary/bin/pipeline.ts
@@ -1,16 +1,16 @@
 #!/usr/bin/env node
 import * as cdk from 'aws-cdk-lib';
 import { PipelineStack } from '../lib/pipeline-stack';
-import { LambdaStack } from '../lib/lambda-stack';
+import { CanaryStack } from '../lib/canary-stack';
 
 const REPO_NAME = "https://github.com/filecoin-project/bacalhau"
 
 const app = new cdk.App();
 
-const lambdaStack = new LambdaStack(app, 'BacalhauCanaryLambda');
+const canaryStack = new CanaryStack(app, 'BacalhauCanary');
 
 new PipelineStack(app, 'BacalhauCanaryPipeline', {
-    lambdaCode: lambdaStack.lambdaCode,
+    lambdaCode: canaryStack.lambdaCode,
     repositoryName: REPO_NAME,
 });
